refactor(selector4nix): rename internal modules
